api: add new callback interface while keeping backward compatibility
This patch adds nfct_callback_register2() and nfct_callback_unregister2()
that allows to register a callback function with a new callback interface
that includes the Netlink message. This fixes an early design error.

This is not nice but it is the only way to resolve this problem without
breaking backward (I don't like function versioning, it is messy).

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>