ct: fix regression with helpers for Linux kernel >= 2.6.34
Since Linux kernel 2.6.34, the attribute validation for CTA_HELP_NAME
requires that the string must be NULL terminated. I think that this
should be fixed in the kernel instead since it breaks old binaries of
the library. However, we're already in 2.6.36-rc, so let's fix it
in user-space and hope that everyone upgrades.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>