expect: export already implement low-level functions
This patch exports several low-level function that allow to build
and parse netlink messages that contain ctnetlink expectation
information. They have been in the tree for quite so long, but
they were not exported. To leave the library in a consistent state
(ctnetlink conntrack functions provide the similar functionality)
I have decided to export them.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>