src: replace CTA_PROTOINFO_DCCP_SEQ by CTA_PROTOINFO_DCCP_HANDSHAKE_SEQ
This patch uses CTA_PROTOINFO_DCCP_HANDSHAKE_SEQ instead which is the
name that is used in the Linux kernel header. Thus, both the headers
and the internal copy for the library are in sync.

This problem was probably introduced at the time that we added support
for the DCCP handshake sequence number.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>