conntrack: add new API to build/parse ctnetlink messages using libmnl
This patch adds support to build and to parse netlink messages
from/to one user-space nf_conntrack object. It uses libmnl, thus
libnetfilter_conntrack now depends on this library.

This is the first patch in the direction of removing the dependency
on the veteran libnfnetlink.

I have decided to update LIBVERSION in this patch. I know it's
recommended to do this before releasing the software. I prefer to
do this so snapshot packages get the correct LIBVERSION.

Signed-off-by: Pablo Neira Ayuso <pablo@gnumonks.org>